Add experimental code on evaluating Char RNN
The experimental code covers the evaluation of Char RNN 
example available in CNTK repository. The code on the CNTK 
repo is written in Python and covers both the training and 
the evaluation of the model.

This C# port covers only the evaluation of the model using the 
model trained through Python.

Signed-off-by: Albert Richard Sanyoto <albertalrisa@gmail.com>