Improve training performance and bug fix
This commit fixes the GetData function, where the substring is
incorrectly applied to the original data instead of the input string.

In addition, the training performance is improved lots by adding one as
the second parameter of TrainingParameterScheduleDouble(). The current
training managed to reach a cross-entropy loss value of around 0.5 in
one epoch and under 1000 minibatch. This is a huge change compared to
the previous iteration where even after 20+ epochs the model is
practically unusable.

Signed-off-by: Albert Richard Sanyoto <albertalrisa@gmail.com>